Added a new study project
This time, I created a website for a travel agency. I used only HTML and learned a little about SEO and how meta descriptions affect it. I also learned about the role of Open Graph tags, why they should be used, and how they can help improve SEO on social media. I understood the purpose of div elements, classes, and IDs, the purpose of HTML entities, and the importance of JavaScript on a web page.